Added get-started, expanding texts, chanterelle vectors, animation of the excellence section block
Made the following changes: 1. Aligned the get-started block with the form and made it responsive. 2. Added all the necessary texts that expand or collapse when the corresponding buttons are pressed. 3. Added fox vectors. 4. Made an animation from 0 to 100% in the excellence section. 5. Improved adaptability so that it also displays well on all types of devices (computers, tablets, phones)